\name{dpeaqms.model}
\docType{data}
\alias{dpeaqms.model}

\title{Differential Protein Expression Model}
\description{The differential protein expression model in JAGS
  format developed in Jow et. al 2011.}
\usage{dpeaqms.model}
\format{A string containing a description of the model in JAGS format}
\source{Bayesian identification of protein differential expression in
  multi-group isobaric labelled mass spectrometry data, Appendix}
\references{Jow et. al 2011}
\keyword{JAGS model}

