\name{dpeaqms.simulatedDataset1}
\docType{data}
\alias{dpeaqms.simulatedDataset1}

\title{Simulated MS/MS dataset}
\description{Simulated Isobaric MS/MS dataset used in Jow et. al. 2011}
\usage{dpeaqms.simulatedDataset1}
\format{Table of values in tsv format}
\source{Bayesian identification of protein differential expression in
  multi-group isobaric labelled mass spectrometry data, Section ??}
\references{Jow et. al 2011}
\keyword{datasets}

